(#1353) Add forward declaration of API methods to a implementation class
Add a Terminal class that receives a message to forward an API method to
a specific implementation method that is saved in a method config store
(called luggage).

Add a Gate class that provides the opportunity the forward an
implementation of a API method to a class or module. All classes that
implements a API method have to include the Seatbelt::Gate module.